% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CIF}
\alias{CIF}
\title{Nonparametric Estimated Cumulative Incidence Function}
\usage{
CIF(time, status12, status13, status14)
}
\arguments{
\item{time}{time}

\item{status12}{indicator history of 1->2 transition}

\item{status13}{indicator history of 1->3 transition}

\item{status14}{indicator history of 1->4 transition}
}
\value{
A data frame contains estimated cumulative incidence function
}
\description{
Nonparametric Estimated Cumulative Incidence Function
}
\examples{
library(dplyr)
library(ggplot2)
Data = data_generation()


status12 = Data$status12
status13 = Data$status12

Data_new = Data \%>\% mutate(status14_new = status14+status04) \%>\% select(status14_new)
status14 = Data_new$status14_new
Data_results = CIF(Data$t1, status12, status13, status14)
ggplot(data=Data_results)+geom_point(aes(t, CIF1))
### plot required...
}
