% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{data_generation}
\alias{data_generation}
\title{Data Generation Function for a Illness-Death model with Masked Causes}
\usage{
data_generation(
  n = 200,
  alpha12 = 2.5,
  beta12 = 2,
  alpha13 = 5,
  beta13 = 2.5,
  lambda23 = 1,
  k23 = 2,
  lambda04 = 4,
  k04 = 2,
  a14 = 1,
  b14 = 4,
  tau = 0.3,
  p = 0.7,
  Pg_mb = 0.3,
  Pg_mc = 0.2,
  C_max = 10,
  tau_0 = 2
)
}
\arguments{
\item{n}{number of observations.}

\item{alpha12}{log-logstic scale parameter; 1->2 transition.}

\item{beta12}{log-logstic shape parameter; 1->2 transition.}

\item{alpha13}{log-logstic scale parameter; 1->3 transition.}

\item{beta13}{log-logstic shape parameter; 1->3 transition.}

\item{lambda23}{Weibull scale parameter; 2->3 transition.}

\item{k23}{Weibull shape parameter; 2->3 transition.}

\item{lambda04}{Weibull scale parameter; 0->4 transition.}

\item{k04}{Weibull shape parameter; 0->4 transition.}

\item{a14}{Uniform lower limit; 1->4 transition with non cured proportion.}

\item{b14}{Uniform upper limit; 1->4 transition with non cured proportion.}

\item{tau}{Pre-specified upper threshold value for cured individual.}

\item{p}{Probability of not-being cured.}

\item{Pg_mb}{Probability of masked given that it was 1->3 transition.}

\item{Pg_mc}{Probability of masked given that it was 1->4 transition.}

\item{C_max}{Maximum value of censoring time.}

\item{tau_0}{Pre-specified lower threshold value for not being cured individual.}
}
\value{
Data frame
}
\description{
Data Generation Function for a Illness-Death model with Masked Causes
}
\examples{
require(dplyr)
Data_temp = data_generation(n=200, alpha12=2.5, beta12=2, alpha13=5, beta13=2.5, lambda23=1, k23=2, lambda04=10, k04=4, a14=4, b14=2, tau=0.3, p=0.7, Pg_mb=0.3, Pg_mc=0.2, C_max=10, tau_0=2)
head(Data_temp)


}
