% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_ver2.R
\name{data_generation_ver2}
\alias{data_generation_ver2}
\title{Data Generation Function for a Illness-Death model with Masked Causes}
\usage{
data_generation_ver2(
  n = 200,
  alpha12 = 3,
  beta12 = 4,
  alpha13 = 3.5,
  beta13 = 3,
  lambda23 = 2.5,
  k23 = 1.5,
  lambda04 = 7,
  k04 = 4,
  a14 = 3,
  b14 = 8,
  p = 0.7,
  C_max = 10,
  tau = 0.3
)
}
\arguments{
\item{n}{number of observations.}

\item{alpha12}{log-logstic scale parameter; 1->2 transition.}

\item{beta12}{log-logstic shape parameter; 1->2 transition.}

\item{alpha13}{log-logstic scale parameter; 1->3 transition.}

\item{beta13}{log-logstic shape parameter; 1->3 transition.}

\item{lambda23}{Weibull scale parameter; 2->3 transition.}

\item{k23}{Weibull shape parameter; 2->3 transition.}

\item{lambda04}{Weibull scale parameter; 0->4 transition.}

\item{k04}{Weibull shape parameter; 0->4 transition.}

\item{a14}{Uniform lower limit; 1->4 transition with non cured proportion.}

\item{b14}{Uniform upper limit; 1->4 transition with non cured proportion.}

\item{p}{Probability of not-being cured.}

\item{C_max}{Maximum value of censoring time.}
}
\value{
Data frame
}
\description{
Data Generation Function for a Illness-Death model with Masked Causes
}
\examples{
require(dplyr)
Data_temp = data_generation_ver2(a14=2,b14=6)
head(Data_temp)
Data_temp \%>\% select(t1, t2, status12, status23, status13, status14) \%>\% filter(status12==1)
Data_temp \%>\% select(t1, t2, status12, status23, status13, status14) \%>\% filter(status13==1)
Data_temp \%>\% select(t1, t2, status12, status23, status13, status14) \%>\% filter(status14==1)
Data_temp \%>\% select(t1, t2, status12, status23, status13, status14) \%>\% filter(status12==1)\%>\%summarise(n=n(), mean(t1), min(t1), max(t1))
Data_temp \%>\% select(t1, t2, status12, status23, status13, status14) \%>\% filter(status13==1)\%>\%summarise(n=n(), mean(t1), min(t1), max(t1))
Data_temp \%>\% select(t1, t2, status12, status23, status13, status14) \%>\% filter(status14==1)\%>\%summarise(n=n(), mean(t1), min(t1), max(t1))
}
