% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f10j}
\alias{f10j}
\title{Title}
\usage{
f10j(alphajk, betajk, alphajl, betajl, t)
}
\arguments{
\item{alphajk}{a scale parameter for transition j->k.}

\item{betajk}{a shape parameter for transition j->2k.}

\item{alphajl}{a scale parameter for transition j->l.}

\item{betajl}{a shape parameter for transition j->l.}

\item{t}{time to event where t>0.}
}
\value{
a value of hazard
}
\description{
Title
}
\examples{
f10j(1,2,1.3,2.3,5)
}
