% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{hazardUnif}
\alias{hazardUnif}
\title{Uniform Hazard Function}
\usage{
hazardUnif(a14, b14, t)
}
\arguments{
\item{a14}{lower limit of the distribution.}

\item{b14}{upper limit of the distribution.}

\item{t}{vector of quntiles.}
}
\value{
hazard function of uniform distribution.
}
\description{
Uniform Hazard Function
}
\examples{
hazardUnif(1,4,2)
}
