% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{hazardWei}
\alias{hazardWei}
\title{Weibull hazard function}
\usage{
hazardWei(k, lambda, t)
}
\arguments{
\item{k}{is a shape parameter.}

\item{lambda}{is a scale parameter.}

\item{t}{is time to event where t>0.}
}
\value{
haaard function of t.
}
\description{
Weibull hazard function
}
\examples{
hazardWei(1,2,1.3)
\dontrun{
hazardWei(1,2,4)
}
}
