% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_like_func.R
\name{log_like_func}
\alias{log_like_func}
\title{log-likelihood function without masked data}
\usage{
log_like_func(p)
}
\arguments{
\item{p}{}
}
\description{
log-likelihood function without masked data
}
