% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_like_func.R
\name{make.NegLogLik}
\alias{make.NegLogLik}
\title{Constructor function that creates a negative log-likelihood function with unmaksed data}
\usage{
make.NegLogLik(
  data,
  fixed = c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE)
)
}
\arguments{
\item{data}{ganerated data}

\item{fixed}{parameters}
}
\value{
negative log-likelihood function
}
\description{
Constructor function that creates a negative log-likelihood function with unmaksed data
}
\examples{
Data <- data_generation_ver1(n=n, alpha12 = alpha12, beta12 = beta12, alpha13 = alpha13, beta13 = beta13, alpha23 = alpha23, beta23 = beta23, tau = tau, p = p, C_max = C_max, alpha04 = alpha04, beta04 = beta04)
nLL <- make.NegLogLik(Data)
phi <- 2/(1-tau)-2 # parameter of the Clayton copula
initial_parms <- c(alpha12, beta12, alpha13, beta13, alpha23, beta23, phi, p)
nlm(nLL, initial_parms)$estimate
}
