% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_like_func.R
\name{make.NegLogLik_first_stage}
\alias{make.NegLogLik_first_stage}
\title{Constructor function that creates a negative log-likelihood function with unmaksed data first stage}
\usage{
make.NegLogLik_first_stage(data, fixed = c(FALSE, FALSE, FALSE, FALSE, FALSE))
}
\arguments{
\item{data}{generated data}

\item{fixed}{parameters}
}
\value{
negative log-likelihood function for first stage
}
\description{
Constructor function that creates a negative log-likelihood function with unmaksed data first stage
}
\examples{
B=10
n=200
alpha12 = 2
beta12 = 4
alpha13 = 3.5
beta13 = 3
alpha23 = 2.5
beta23 = 1.5
tau = 0.3
p = 0.7
C_max = 10
alpha04 = 7
beta04 = 4
Data <- data_generation_ver1(n=n, alpha12 = alpha12, beta12 = beta12, alpha13 = alpha13, beta13 = beta13, alpha23 = alpha23, beta23 = beta23, tau = tau, p = p, C_max = C_max, alpha04 = alpha04, beta04 = beta04)
nLL <- make.NegLogLik_first_stage(Data)
initial_parms <- c(alpha12, beta12, alpha13, beta13, p)
nlm(nLL, initial_parms)$estimate
}
