% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation_ver1.R
\name{data_generation_ver1}
\alias{data_generation_ver1}
\title{Title}
\usage{
data_generation_ver1(
  n = 400,
  alpha12 = 1,
  beta12 = 1,
  alpha13 = 2,
  beta13 = 2,
  alpha23 = 1,
  beta23 = 2,
  tau = 0.2,
  p = 0.7,
  C_max = 10,
  alpha04 = 7,
  beta04 = 4
)
}
\arguments{
\item{n}{number of observations.}

\item{alpha12}{log-logstic scale parameter; 1->2 transition.}

\item{beta12}{log-logstic shape parameter; 1->2 transition.}

\item{alpha13}{log-logstic scale parameter; 1->3 transition.}

\item{beta13}{log-logstic shape parameter; 1->3 transition.}

\item{tau}{Kendall's tau; dependency measure}

\item{p}{Probability of not-being cured.}

\item{C_max}{Maximum value of censoring time.}

\item{lambda23}{Weibull scale parameter; 2->3 transition.}

\item{k23}{Weibull shape parameter; 2->3 transition.}
}
\description{
Title
}
\examples{
require(dplyr)
Data_temp <- data_generation_ver1(n = 1000)
head(Data_temp)
Data_temp \%>\%
  select(t1, t2, status12, status23, status13) \%>\%
  filter(status23 == 1)
}
