% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{Estimation_unmasked}
\alias{Estimation_unmasked}
\title{Estimation unmasked data}
\usage{
Estimation_unmasked(
  n = 200,
  alpha12 = 2,
  beta12 = 4,
  alpha13 = 3.5,
  beta13 = 3,
  alpha23 = 2.5,
  beta23 = 1.5,
  tau = 0.3,
  p = 0.7,
  C_max = 10,
  alpha04 = 7,
  beta04 = 4
)
}
\value{
nlm output with some information of generated data
}
\description{
Estimation unmasked data
}
\examples{
B=10
n=200
alpha12 = 2
beta12 = 4
alpha13 = 3.5
beta13 = 3
alpha23 = 2.5
beta23 = 1.5
tau = 0.3
p = 0.7
C_max = 10
alpha04 = 7
beta04 = 4
Result <- Estimation_unmasked(n, alpha12, beta12, alpha13, beta13, alpha23, beta23, tau, p, C_max, alpha04, beta04)
print(Result)
}
