% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f10j}
\alias{f10j}
\title{Title}
\usage{
f10j(alphaj, betaj, alphak, betak, t)
}
\arguments{
\item{alphaj}{a scale parameter for transition 1->2.}

\item{betaj}{a shape parameter for transition 1->2.}

\item{alphak}{a scale parameter for transition 1->3.}

\item{betak}{a shape parameter for transition 1->3.}

\item{t}{time to event where t>0.}
}
\description{
Title
}
\examples{
f10j(1,2,1.3,2.3,5)
}
