% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazardWei.R
\name{hazardWei}
\alias{hazardWei}
\title{Weibull hazard function}
\usage{
hazardWei(beta23, alpha23, t)
}
\arguments{
\item{beta23}{a shape parameter.}

\item{alpha23}{a scale parameter.}

\item{t}{time to event where t>0.}
}
\value{
hazard function of t.
}
\description{
Weibull hazard function
}
\examples{
hazardWei(1,2,1.3)
\dontrun{
hazardWei(1,2,4)
}
}
