% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{Simulation_unmasked}
\alias{Simulation_unmasked}
\title{Simulation unmasked data}
\usage{
Simulation_unmasked(
  B = 10,
  n = 200,
  alpha12 = 2,
  beta12 = 4,
  alpha13 = 3.5,
  beta13 = 3,
  alpha23 = 2.5,
  beta23 = 1.5,
  tau = 0.3,
  p = 0.7,
  C_max = 10,
  alpha04 = 7,
  beta04 = 4
)
}
\value{
list of estimation outputs from 1:B
}
\description{
Simulation unmasked data
}
\examples{
B=3
n=200
alpha12 = 2
beta12 = 4
alpha13 = 3.5
beta13 = 3
alpha23 = 2.5
beta23 = 1.5
tau = 0.3
p = 0.7
C_max = 10
alpha04 = 7
beta04 = 4
Sim_result <- Simulation_unmasked(B, n, alpha12, beta12, alpha13, beta13, alpha23, beta23, tau, p, C_max, alpha04, beta04)
total_num <- length(Sim_result)
estimations <- c()
for (i in 1:total_num){
 estimation_temp <- Sim_result[[i]]$estimate
 estimations <- rbind(estimations, estimation_temp)
}
print(estimations)
}
